+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 1.x branch
 ## 1.0 branch
+### 1.0.3
+* the target directory is created automatically, if it does not exist;
+* `BackupManager::index()` returns an array of entities;
+* added `MYSQL_BACKUP` constant.
+
 ### 1.0.2
 * methods that have been deprecated with CakePHP 3.4 have been replaced;
 * updated for CakePHP 3.4.